added cositas

+++ src/main/java/aiss/api/resources/LibraryResource.java
@@ -1,7 +1,7 @@
 package aiss.api.resources;
 
 import java.net.URI;
-
+import java.util.ArrayList;
 import java.util.Collection;
 import java.util.Comparator;
 import java.util.stream.Collectors;
@@ -12,22 +12,22 @@ import javax.ws.rs.GET;
 import javax.ws.rs.POST;
 import javax.ws.rs.PUT;
 import javax.ws.rs.Path;
+import javax.ws.rs.PathParam;
 import javax.ws.rs.Produces;
 import javax.ws.rs.QueryParam;
-import javax.ws.rs.PathParam;
 import javax.ws.rs.core.Context;
 import javax.ws.rs.core.Response;
+import javax.ws.rs.core.Response.ResponseBuilder;
 import javax.ws.rs.core.UriBuilder;
 import javax.ws.rs.core.UriInfo;
-import javax.ws.rs.core.Response.ResponseBuilder;
 
 import org.jboss.resteasy.spi.BadRequestException;
 import org.jboss.resteasy.spi.NotFoundException;
 
-import aiss.model.Library;
 import aiss.model.Film;
-import aiss.model.repository.MapLibraryRepository;
+import aiss.model.Library;
 import aiss.model.repository.LibraryRepository;
+import aiss.model.repository.MapLibraryRepository;
 
 
 
@@ -184,17 +184,21 @@ public class LibraryResource {
 	@Path("/{libraryId}/{filmId}")
 	public Response removeFilm(@PathParam("libraryId") String libraryId, @PathParam("filmId") String filmId) {
 		Library library = repository.getLibrary(libraryId);
-		Film film = repository.getFilm(filmId);
-		
-		if (library==null)
-			throw new NotFoundException("The library with id=" + libraryId + " was not found");
-		
-		if (film == null)
-			throw new NotFoundException("The film with id=" + filmId + " was not found");
-		
-		
-		repository.removeFilm(libraryId, filmId);		
-		
+		if(!filmId.equals("*")) {
+			Film film = repository.getFilm(filmId);
+			
+			if (library==null)
+				throw new NotFoundException("The library with id=" + libraryId + " was not found");
+			
+			if (film == null)
+				throw new NotFoundException("The film with id=" + filmId + " was not found");
+			
+			
+			repository.removeFilm(libraryId, filmId);		
+		}else {
+			library.setFilms(new ArrayList<>());
+			repository.updateLibrary(library);
+		}
 		return Response.noContent().build();
 	}
 }
